Move VerfahrenTileSkeleton into /component folder
