+++ app/components/verfahren/VerfahrenList.tsx
@@ -1,4 +1,4 @@
-import VerfahrenTileSkeleton from "~/components/skeletons/VerfahrenTileSkeleton.static";
+import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
 import VerfahrenTile from "~/components/verfahren/VerfahrenTile";
 import { VERFAHREN_SKELETONS } from "~/config/verfahrenSkeletons";
 import { Verfahren } from "~/routes/verfahren/_index";
