+++ app/routes/verfahren/_index.tsx
@@ -7,10 +7,10 @@ import { useParamsState } from "~/components/hooks/useParamsState";
 import InputSelect from "~/components/InputSelect";
 import ScrollToTopButton from "~/components/ScrollToTopButton";
 import Search from "~/components/Search";
-import VerfahrenTileSkeleton from "~/components/skeletons/VerfahrenTileSkeleton.static";
 import { VerfahrenCounter } from "~/components/verfahren/VerfahrenCounter";
 import { VerfahrenList } from "~/components/verfahren/VerfahrenList";
 import { VerfahrenLoadMoreButton } from "~/components/verfahren/VerfahrenLoadMoreButton";
+import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
 import { sortOptions, VERFAHREN_PAGE_LIMIT } from "~/config/verfahren";
 import { VERFAHREN_SKELETONS } from "~/config/verfahrenSkeletons";
 import { authContext } from "~/middleware/auth.server";
